www: Extract BuildersTable component
